refactor: constants 위치 이동
